feat(vision-manifold): saliency refinement, multi-horizon tracking, integration test, benchmark, demo wiring
Sub-crate (59 tests, 0 clippy warnings):
- refine_from_attention(): closed-loop active inference — surprise map drives
  contrastive encoder weight updates toward salient scene regions
- evaluate_horizons() + HorizonAccuracy: 4-horizon prediction accuracy
  (0.033s/0.1s/0.5s/1.0s) with serde support for telemetry export
- camera.rs: replaced nokhwa with v4l crate, fixed CameraManifold type mismatch
- HorizonAccuracy re-exported from lib.rs

Integration (behind vision-manifold feature):
- tests/vision_manifold_integration.rs: 7 tests — VisionBridge→cycle_with_hv,
  multi-frame pipeline, scene change propagation, CameraManifold convenience,
  telemetry consistency, horizon accuracy, saliency refinement
- examples/vision_manifold_benchmark.rs: 6 benchmarks — static convergence,
  scene change detection, alternating pattern learning, multi-scale
  discrimination, temporal prediction accuracy, throughput

Demo UI wiring (api_module + vision-manifold features):
- DemoCycleData: 10 vision telemetry fields (prediction_error, coherence,
  attention_entropy, salient_patches, horizon_errors, encode/evolve timing)
- DemoRunner: enable_vision()/disable_vision() with CameraManifold integration
- WebSocket: {"vision": true/false} client message for toggling vision mode

+++ symthaea/src/cognitive_loop/accessors/behavior.rs
@@ -222,7 +222,7 @@ impl CognitiveLoopService {
         let (_, pattern_confidence) = self.temporal_signature_encoder.classify_state();
         let consciousness_level =
             super::super::snapshot::ConsciousnessSnapshot::compute_consciousness_level(
-                self.prediction_confidence,
+                self.prediction_confidence as f32,
                 self.coherence_bridge.smoothed_coherence(),
                 self.flow_state.intensity,
                 pattern_confidence,
